fix(dashboard): un utilisateur "entreprise" peut voir toutes ses fiscalités minières sur son dashbaord fix#322
